+++ src/Presentation/Console/Application/ApplicationFactory.php
@@ -38,6 +38,9 @@ use ArtARTs36\MergeRequestLinter\Infrastructure\Notifications\Notifier\MessageCr
 use ArtARTs36\MergeRequestLinter\Infrastructure\Notifications\Notifier\NotifierFactory;
 use ArtARTs36\MergeRequestLinter\Infrastructure\Text\Renderer\TwigRenderer;
 use ArtARTs36\MergeRequestLinter\Infrastructure\ToolInfo\ToolInfoFactory;
+use ArtARTs36\MergeRequestLinter\Presentation\Console\Application\Providers\NotificationsProvider;
+use ArtARTs36\MergeRequestLinter\Presentation\Console\Application\Providers\RuleProvider;
+use ArtARTs36\MergeRequestLinter\Presentation\Console\Application\Providers\ServiceProvider;
 use ArtARTs36\MergeRequestLinter\Presentation\Console\Command\DumpCommand;
 use ArtARTs36\MergeRequestLinter\Presentation\Console\Command\InfoCommand;
 use ArtARTs36\MergeRequestLinter\Presentation\Console\Command\InstallCommand;
@@ -59,6 +62,11 @@ use Symfony\Component\Console\Output\OutputInterface;
 
 class ApplicationFactory
 {
+    private const PROVIDERS = [
+        NotificationsProvider::class,
+        RuleProvider::class,
+    ];
+
     public function __construct(
         private readonly MapContainer $container = new MapContainer(),
         private readonly Environment $environment = new LocalEnvironment(),
@@ -111,7 +119,7 @@ class ApplicationFactory
 
         $events = $this->registerEventDispatcher();
 
-        $this->registerNotifications();
+        $this->runProviders();
 
         $application = new Application($metrics);
 
@@ -126,6 +134,14 @@ class ApplicationFactory
         $application->add(new InfoCommand(new ShowToolInfoHandler(new ToolInfoFactory($clock))));
 
         return $application;
+    }
+
+    private function runProviders(): void
+    {
+        /** @var class-string<ServiceProvider> $providerClass */
+        foreach (self::PROVIDERS as $providerClass) {
+            (new $providerClass($this->container))->provide();
+        }
     }
 
     /**
@@ -149,46 +165,6 @@ class ApplicationFactory
         $this->container->set(Clock::class, $clock);
 
         return $clock;
-    }
-
-    private function registerNotifications(): void
-    {
-        $notificationsListener = function (ConfigResolvedEvent $event) {
-            $this
-                ->createNotificationsListenerRegistrar($event->config->config)
-                ->register($this->container->get(EventManager::class));
-        };
-
-        $this
-            ->container
-            ->get(EventManager::class)
-            ->listen(ConfigResolvedEvent::class, new CallbackListener(
-                'registration notifications',
-                $notificationsListener,
-            ));
-    }
-
-    private function createNotificationsListenerRegistrar(Config $config): ListenerRegistrar
-    {
-        $logger = $this->container->get(ContextLogger::class);
-
-        $notifier = (new NotifierFactory(
-            $this->container->get(HttpClientFactory::class)->create(
-                $config->getHttpClient()
-            ),
-            $this->container->get(ClockInterface::class),
-            $logger,
-        ))->create();
-
-        return new ListenerRegistrar(
-            $config->getNotifications(),
-            new ListenerFactory(
-                $notifier,
-                $this->container->get(OperatorResolver::class),
-                new MessageCreator(TwigRenderer::create()),
-                $logger,
-            ),
-        );
     }
 
     private function registerHttpClientFactory(): ClientFactory
