+++ src/Presentation/Console/Application/Providers/ServiceProvider.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace ArtARTs36\MergeRequestLinter\Presentation\Console\Application\Providers;
+
+interface ServiceProvider
+{
+    public function provide(): void;
+}
